Allow exact value of timestamp to be specified
Previously, the logger self established the timestamp. However, the actual occurrence of a logged message might have been some time in the past. For instance, due to internal queueing or that a log message could not be registered earlier. This makes searching the log messages in CloudWatch complex when there are thousands of messages per minute.