Fix messages getting lost.
The logic of AddSingleMessage is unclear from the logic itself. In the previous version, the contents of the message would be replaced always be an error when the queue size is exceeded and the new text was being logged once every five minutes. In all other cases, the message would be lost without a sign.