Update AWSLoggerCore.cs
The logic of AddSingleMessage is unclear from the logic itself. In the previous version, the contents of the message would be replaced always be an error when the queue size is exceeded and the new text was being logged once every five minutes. In all other cases, no message was logged. In all cases, the original cases would be lost without displaying the amount lost.

The reason why messages are not being processed remains to be studied. It occurs despite sufficient sizing and several times a day in a large SAAS-environment.

Also adds a timestamp to messages to allow deviating timestamps.